update userAnexo endpoints

+++ productengineering/domain/src/main/java/br/com/todeschini/domain/business/auth/authservice/AuthServiceImpl.java
@@ -20,7 +20,7 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public void validateSelfOrAdmin(Long userId) {
+    public void validateSelfOrAdmin(Integer userId) {
         authMethods.validateSelfOrAdmin(userId);
     }
 }
